added income-needed section to retirement

+++ Shared/Models/FamilyData/Person.cs
@@ -28,7 +28,7 @@ public class Person {
     public int? Age { get; set; }
     public int? RetirementAge { get; set; }
     public int? SSAge { get; set; }
-    public int? SSAnnual { get; set; }
+    public double? SSAnnual { get; set; }
     public string? Identifier { get; set; }
     public string? Employer { get; set; }
     public EmployerPlan EmployerPlan { get; set; }
